# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gouda Giggles Charcuterie API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  serves?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  featured?: boolean;
}

export interface OrderItem {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  eventDate: string;
  deliveryAddress: string;
  /** @nullable */
  specialInstructions?: string | null;
  items: OrderItem[];
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  eventDate: string;
  deliveryAddress: string;
  /** @nullable */
  specialInstructions?: string | null;
  status: string;
  total: number;
  items: OrderItem[];
  createdAt: string;
}

export interface PaymentInput {
  sourceId: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  eventDate: string;
  deliveryAddress: string;
  /** @nullable */
  specialInstructions?: string | null;
  items: OrderItem[];
}

export interface QuoteInput {
  name: string;
  email: string;
  phone: string;
  eventType: string;
  eventDate: string;
  guestCount: number;
  message: string;
}

export interface Quote {
  id: number;
  name: string;
  email: string;
  phone: string;
  eventType: string;
  eventDate: string;
  guestCount: number;
  message: string;
  createdAt: string;
}

export interface GalleryImage {
  id: number;
  url: string;
  caption: string;
  category: string;
}

export interface StoreStats {
  totalProducts: number;
  categories: number;
  startingPrice: number;
  happyCustomers?: number;
}

export type ListProductsParams = {
category?: string;
};