# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gouda Giggles Charcuterie API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    post:
      operationId: createPayment
      tags: [orders]
      summary: Process Square payment and create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment successful, order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
        "402":
          description: Payment failed
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
...
    Order:
      type: object
      required: [id, customerName, customerEmail, customerPhone, eventDate, deliveryAddress, status, total, items, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        eventDate:
          type: string
        deliveryAddress:
          type: string
        specialInstructions:
          type: ["string", "null"]
        status:
          type: string
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [sourceId, customerName, customerEmail, customerPhone, eventDate, deliveryAddress, items]
      properties:
        sourceId:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        eventDate:
          type: string
        deliveryAddress:
          type: string
        specialInstructions:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
...